Mocked out more of the scope
The Scope class now communicates with the mocked serial device in the manner
stated in the spec. Data is streamed back properly, but the mock is very much
synchronous. Since the messages coming from the mock device are serialized (and
therefore ideal), we may need to implement things differently for the real
hardware.